Refactor multipart streaming
